eviction type corrected

+++ src/io/cache/prefetching_cache.cpp
@@ -1262,8 +1262,10 @@ void prefetching_cache::evict_loop(const std::stop_token& st)
       _cfg.dispose_on_idle || eviction_requested || _pool->should_start_evicting();
     if (!should_evict) { continue; }
 
-    size_t const need = _cfg.dispose_on_idle ? std::numeric_limits<size_t>::max()
-                                             : _pool->total_allocated_chunks() * 0.25;
+    size_t const need =
+      _cfg.dispose_on_idle
+        ? std::numeric_limits<size_t>::max()
+        : static_cast<size_t>(static_cast<double>(_pool->total_allocated_chunks()) * 0.25);
 
     for (auto& [_, buffers] : reclaim_by_numa) {
       buffers.clear();
